New version of blazorwasm
