--- src/Content/blazorwasm/BlazorApp/Client/Pages/PhpGateway.razor
@@ -1,17 +1,0 @@
-﻿@page "/php/{*sth}"
-@using Peachpie.Blazor
-
-<PhpScriptProvider ContextLifetime="@SessionLifetime.OnNavigationChanged" Type="@PhpScriptProviderType.ScriptProvider">
-    <Navigating>
-        <p>Navigating to script</p>
-    </Navigating>
-    <NotFound>
-        <p>Script not found</p>
-    </NotFound>
-</PhpScriptProvider>
-
-@code
-{
-    [Parameter]
-    public string sth { get; set; }
-}
