--- src/Content/blazorwasm/BlazorApp/Client/Pages/PhpScript.razor
@@ -1,11 +1,0 @@
-﻿@page "/phpstandalone.php"
-@using Peachpie.Blazor
-
-<PhpScriptProvider ContextLifetime="@SessionLifetime.Persistant" Type="@PhpScriptProviderType.Script" ScriptName="fileManagment/index.php">
-    <Navigating>
-        <p>Navigating</p>
-    </Navigating>
-    <NotFound>
-        <p>Script not found</p>
-    </NotFound>
-</PhpScriptProvider>
