+++ src/Content/blazorwasm/Client/App.razor
@@ -1,0 +1,10 @@
+@namespace MyPHPWebsite.1.Client
+
+<Microsoft.AspNetCore.Components.Routing.Router AppAssembly="@typeof(Program).Assembly" PreferExactMatches="@true">
+    <Found Context="routeData">
+        <RouteView RouteData="@routeData"/>
+    </Found>
+    <NotFound>
+        <p>Sorry, there's nothing at this address.</p>
+    </NotFound>
+</Microsoft.AspNetCore.Components.Routing.Router>
