+++ src/Content/blazorwasm/Client/BlazorApp.Client.csproj
@@ -8,11 +8,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="5.0.3" />
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="5.0.3" PrivateAssets="all" />
     <PackageReference Include="System.Net.Http.Json" Version="5.0.0" />
-	<PackageReference Include="Peachpie.Blazor" Version="1.0.12087"/>
+	  <PackageReference Include="Peachpie.Blazor" Version="1.0.12087"/>
+    <ProjectReference Include="..\PHPScripts\PHPScripts.msbuildproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\PHPScripts\PHPScripts.msbuildproj" />
-  </ItemGroup>
-
+  
 </Project>
