+++ src/Content/blazorwasm/Client/Index.razor
@@ -1,9 +1,7 @@
 ﻿@page "/"
 @using Peachpie.Blazor
 
-<h1>Hello, world!</h1>
-
-Welcome to your new app.
+<h1>Hello, world from C#!</h1>
 
 <PhpScriptProvider ContextLifetime="@SessionLifetime.OnNavigationChanged" Type="@PhpScriptProviderType.Script" ScriptName="index.php">
     <Navigating>
