+++ src/Content/blazorwasm/Client/Program.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using Peachpie.Blazor;
 using System.Threading.Tasks;
 
-namespace BlazorApp.Client
+namespace MyPHPWebsite.1.Client
 {
     public class Program
     {
