+++ src/Content/blazorwasm/Client/wwwroot/index.html
@@ -1,0 +1,17 @@
+﻿<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
+    <title>PeachPieBlazorApp</title>
+    <base href="/" />
+</head>
+
+<body>
+    <div id="app">Loading...</div>
+
+    <script src="_framework/blazor.webassembly.js"></script>
+</body>
+
+</html>
