--- src/Content/blazorwasm/PHPScripts/fileManagment/chooseForm.php
@@ -1,7 +1,0 @@
-﻿<form id="form1" action="phpstandalone.php" method="get">
-    <label for="action1">Generate graph: </label><br/>
-    <input id="action1" name="action" type="radio" value="generate"/><br/>
-    <label for="action2">Load graph: </label><br/>
-    <input id="action2" name="action" type="radio" value="inputFile"/><br/>
-    <input type="submit" value="Submit"/>
-</form>
