--- src/Content/blazorwasm/PHPScripts/fileManagment/downloadForm.php
@@ -1,4 +1,0 @@
-﻿<form action="phpstandalone.php" method="get">
-	<input type="hidden" name="action" value="download"/><br/>
-    <input type="submit" value="Download"/><br/>
-</form>
