--- src/Content/blazorwasm/PHPScripts/fileManagment/fileForm.php
@@ -1,6 +1,0 @@
-﻿<form action="phpstandalone.php" method="get">
-    <label for="file">CSV file of graph: </label><br/>
-	<input type="file" id="file" name="file1"/><br/>
-    <input type="hidden" name="action" value="parametersSetF"/><br/>
-    <input type="submit" value="Submit"/><br/>
-</form>
