--- src/Content/blazorwasm/PHPScripts/php/example3.php
@@ -1,8 +1,0 @@
-﻿<h2>Form example 3</h2>
-<p>You can upload files and their content</p>
-
-<form id="form4" action="php/viewform.php" method="get">
-	<input type="file" name="file1"/>
-    <input type="file" name="file2"/>
-	<input type="submit" value="Submit"/>
-</form>
