+++ src/Content/blazorwasm/Server/BlazorApp.Server.csproj
@@ -7,11 +7,7 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="5.0.3" />
 	  <PackageReference Include="Peachpie.Blazor" Version="1.0.12087"/>
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\Client\BlazorApp.Client.csproj" />
   </ItemGroup>
-
 
 </Project>
